Create hw1_ex.c
 You are to write three functions that print numbers in sequence. The sequence may be increasing or decreasing. The first function, do_for, will use a for loop; the second, do_while, will use a while loop, and the third, do_do, will use a do … while loop.
Each function takes 2 arguments; call them m and n. m is the starting number; the n is the ending number.
If m < n, then print m, m+1, …, n, each number on a separate line.
If m = n, then just print m.
If m > n, then print m, m-1, …, n, each number on a separate line.

Each function must have a third argument.  This argument is to be the increment or decrement. It must be positive. So, for example, if the function do_for is called as do_for(3, 10, 2), then it would print “3, 5, 7, 9” (without the quotes), and if it is called as do_for(10, 3, 2), then it would print $10, 8, 6, 4”. If it is called as do_for(3, 10, -2), the function would print Error: invalid increment -2.