Create permute.c
Write a recursive function that prints all permutations of the characters in a string. Your program is to prompt the user with “> ” (note the blank space after the “>”). The user then enters a string of letters or digits or both, and the program lists the permutations. It then reprompts the user for a new string. It continues doing this until the user enters either a blank line or types an end-of-file (control-D as the first character).
Here is an example. The “%” is a shell prompt. The “↵” means the user typed return or enter (or otherwise entered a newline). Your listing of the permutations need not be in the same order, but it must include all permutations of the input.
> 123↵
123
231
312
132
213
321
> ab↵
ab
ba
> ↵