Create life.c
This is the first part of the famous Game of Life program. It uses a board that is a two-dimensional rectangle divided into small, equal-sized squares. Each square may be populated or unpopulated. The neighbors of a cell are the cells that adjoin it vertically, horizontally, or diagonally.
At each step, the following transformations are applied to all the cells.
If a populated cell adjoins less than 2 other populated cells, or more than 3 populated cells, it becomes unpopulated.
If a populated cell adjoins 2 or 3 populated cells, it remains populated.
If an unpopulated cell adjoins exactly 3 populated cells, it becomes populated.
As a first step, you will read the dimensions of the board from the standard input. This will be given as two integers separated by an “x” (the 24th letter of the alphabet) and possibly white space before and after the numbers and the “x”. Your output is to be a square composed of “-”es and “|”es. If the numbers read are “2 x 8”, your square should enclose a 2×8 rectangle. The inside of the rectangle is to be blank.