Update life2.c
This is the second part of the Game of Life problem. The next step is to load a predefined pattern onto the board for the game. The predefined pattern is in a file. That file contains rows of “X”s and “O”s (the characters “X” and “O”). The “X” corresponds to a populated cell and the “O” to an unpopulated one. Each line in that file should have the same number of columns. Your program is to read the file, load the configuration into memory, and print it out, properly framed by the board, and stop.

Your program will take the file with the predetermined pattern as its only argument. It is to print out the board. Note your program will have to count the rows and columns as these will not be given. Report any errors such as a line too long or too short.

As an example, the input file
XOXXOOXOX
OOOXXXOOO
OXOXOXOXO
XOXOXOXOX
defines a board of 4 rows and 9 columns. It is a valid board. But the board
OXOXO
XXXX
OOOOOO
is not valid as the rows are of different lengths, and the board
XOXXOOXOX
OOOXYXOOO
OXOXOXOXO
XOXYXOXOX
is not valid because there is a character other than X, O, and newline in the file (specifically, two instances of “Y”). Your program is to report errors in these cases, and then terminate.
The board dimensions will never be more than 100 rows and 100 columns.