Create lookups.c
Write a program to search for words in a file. Your program is to work like this.
First, read in a file of no more than 1000 words (called a dictionary), and store them in a list. This is to be a list of character pointers, each pointing to a word. You will have to use malloc(3), which allocates space dynamically. Each line, including the last, will be ended with a newline, and no line will be over 1024 characters.

Then, read a second file. This file is also a list of words. Search your list for each word, and for each word print the number of words checked before you either get a match or find it is not in the list. For letters, the case must match (so “N” does not match “n”). Also, after you have finished searching for the words, print the average number of words checked.

The two files are given as command-line arguments to your program. The first argument is the dictionary, and the second is the list of words to be found.

Once you have read in the word list from the dictionary, sort the words. Then proceed as in problem 3. So, you will read a second file, search the list for each word in the file, and for each word print the number of words checked before you either get a match or find it is not in the list. Also, after you have finished searching for the words, print the average number of words checked.