Create life3.c
This is the third step in the Game of Life. The next step is to transform the board into the next generation. Each tranformation is a step. Recall the rules
are as follows. Each square on the board may be populated (indicated by an ‘O’) or unpopulated (indicated by an
‘X’). The neighbors of a cell are the cells that adjoin it vertically, horizontally, or diagonally. At each step, the
following transformations are applied to all the cells.
(a) If a populated cell adjoins less than 2 other populated cells, or more than 3 populated cells, it becomes unpopulated.
(b) If a populated cell adjoins 2 or 3 populated cells, it remains populated.
(c) If an unpopulated cell adjoins exactly 3 populated cells, it becomes populated.
Your program will take the file with the predetermined pattern as its only argument. It is to print out the initial
board (generation 0), and run for 50 generations (steps), printing out the board every 5 generations (so it should
print generations 0, 5, 10, . . . , 50).

As before, the board dimensions will never be more than 100 rows and 100 columns.
The output is to be written to the standard output. If there are no problems, your program is to return an exit
status code of 0; otherwise, it is to return an exit status code of 1. Successful output looks as above. The following
are the error messages:
• If malloc(3) fails, use perror(3) to print the message. If, for example, the return value of malloc is stored in p, then you would do the test as:
if (p == NULL) perror(s);
...
where p holds the return value of malloc and s is the string you are saving in the allocated space.
• If there are not exactly 2 command-line arguments (the program name and the file containing the pattern for
the board), use an fprintf of the following form:
fprintf(stderr, "Usage: %s board_pattern\n", argv[0]);
• If opening the file fails, use perror to print an error message, like this:
perror(fname); where fname is the name of the file that you were trying to open.
• If a line in the file is of the wrong length, use an fprintf of the following form: fprintf(stderr, "%s: line %d is wrong length\n", fname, lineno); where fname is the name of the file and lineno is the line number of the bad line.
• If a line in the file contains an invalid character, use an fprintf of the following form: fprintf(stderr, "%s: bad character ’%c’ in line%d\n", fname, ch, lineno); where fname is the name of the file, ch is the invalid character, and lineno is the line number of the bad line