Create life3ex.c
Addition to life3.c. Added command line options:
• -n followed by an integer (with or without intervening space): number of generations to run the program for
• -p followed by an integer (with or without intervening space): print the board at intervals of this many
The integers must be positive. For example, running
life3ex -n 100 -p 10 board
would begin with the pattern in board, run it for 100 generations, printing every 10th one (that is, it would print
generations 0, 10, 20, . . . , 10, where generation 0 is the initial configuration). Always print the initial and final
configurations.
In addition to everything for the program in life3.c, if either option is not followed by a number, use an
fprintf of the following form:
fprintf(stderr, "%s: %c must be followed by a positive number\n", fname, option);
where fname is the name of the file and option the option involved, and if a bad option is given, use an fprintf
of the following form:
fprintf(stderr, "%s: %c: invalid option\n", fname, option); where fname and option are as above.